refactor: main 변경사항 반영, 아이콘 수정
