+++ src/app/user/layout.tsx
@@ -1,4 +1,4 @@
-import { IcnSquare } from "@/asset/svg";
+import { IcnMy, IcnPlus, IcnSearch, IcnSquare } from "@/asset/svg";
 import NavBar from "@/shared/component/NavBar";
 import type { Metadata } from "next";
 
@@ -23,21 +23,21 @@ export default function UserLayout({
           대시보드
         </NavBar.Item>
         <NavBar.Item
-          icon={<IcnSquare width={24} height={24} />}
+          icon={<IcnPlus width={24} height={24} />}
           mode="stroke"
           href="/user/create-group"
         >
           그룹 만들기
         </NavBar.Item>
         <NavBar.Item
-          icon={<IcnSquare width={24} height={24} />}
+          icon={<IcnSearch width={24} height={24} />}
           mode="stroke"
           href="/user/join-group"
         >
           그룹 가입하기
         </NavBar.Item>
         <NavBar.Item
-          icon={<IcnSquare width={24} height={24} />}
+          icon={<IcnMy width={24} height={24} />}
           mode="stroke"
           href="/user/setting"
         >
