+++ src/view/user/setting/StudyList/index.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { StudyListData } from "@/shared/constant/example";
-import { backPanelStyle, headingStyle, wrapperStyle } from "../index.css";
+import { backPanelStyle, headingStyle, wrapperStyle } from "@/view/user/setting/index.css";
 import { StudyListTableProvider } from "./StudyListTable/StudyListProvider";
 import StudyListTable from "./StudyListTable/StudyListTable";
